"Fix" missing pixels bug, see longer description
So, I don't know of a perfect fix off the top of my head, but using
-1,-1 as the arbitrary point was probably a bad idea and it's just
chance that I almost never (or maybe never?) used a square resolution
till now, when I had an edge down that diagonal. But having triangles
with a shared edge exactly down the diagonal of the viewport is not
exactly uncommon and then it's just 50-50 whether you would run into
this.  For now I just went to -1,-2.5 for the heck of it.

I could change the comparisons to >= 0 instead of > 0 so it would just
draw the pixel twice in those cases and the second triangle wins?

Something to think about.