Testing OpenMP with draw_triangle_fill
Not sure it's worth it, even the best case with 2 large triangles like
the texturing demo only see a ~30% speedup currently.  Even if I could
increase that with fine tuning the pragma options, most scenes have are
made up of many many smaller triangles, so the overhead of creating and
destroying threads for each triangle would keep the speedup to much
less than what we'd want to see.  I need to look at how other projects
do it for comparison.