Thanks to billsix for catching these two issues
billsix is working on getting PortableGL building on Windows here
https://github.com/billsix/PortableGL

As far as the lack of initializing the zero vector in chalfhedge.c, that's
what I get for not being more careful when porting it from rsw_halfedge.cpp.

I really should do more programming in pure C to force myself to really
test and flesh out my C libraries.