Start writing clipping tests...
Just added x and y plane clipping tests for triangles in fill, line, and
point modes, with points and lines in sizes 8 and 32 as well.

Currently triangle clipping in point mode is broken and my thick line
algorithm is flawed, sometimes leaving holes, but I'm putting the output
in expected anyway for now.  For the thick line drawing, I might end up
going back to the simpler and faster method I had before, where the ends
are perpendicular to the direction of greatest change, (ie a line with
slope (-inf, -1] and (1, +infinite) have horizontal ends, while lines between (-1, 1]
have vertical ends.  It looks worse the thicker the line.