Ran perf tests 0.97ish vs current 0.98
In most respects 0.98 with guardband "clipping" for the xy planes is
faster between 3-5x for the clipping tests and 50-70% for the point
tests.  It is 10-20% slower for the triangle tests and 5-10% slower
for the line_perf test. Also noteworthy is that clang is generally
slightly faster.

The slower tests (triangles and lines) are less realistic as there is
no clipping required (maybe I should change that).  In most real world
use cases I think the much faster clipping would balance out the slightly
slower drawing when no clipping was needed.

Still, long term I should look into improving the speed of rasterizing
both lines and triangles.  Points seem to be doing great in 0.98.