init_glContext() now sets the current context
It simplifies things and it is what users would want 99.9% of the time
anyway so it saves them a call to set_glContext() immediately after.

Even in the case where they want to initialize multiple contexts
it really doesn't change anything because they would just call
set_glContext() for the one they want afterward anyway.