More progress on imgui backend, slow going...
Discovered several bugs and issues, subtle things I never really think
about because I never use PGL in a way that exposes them.  This is why porting
real projects is always the best way to test PGL.

The first and most obvious is I never really tested glDrawElements (and
its relatives.  I chose to make the indices pointer a GLsizei because
it was an offset, but I treated it like and index rather than a byte
offset.  I think I'll do what I did with AttribVertexPointer.