Rename EXCLUDE_GLSL to PGL_EXCLUDE_GLSL
Might eventually have the option PGL_MANGLE_GLSL instead of or in
addition to EXCLUDE as a way to prevent clashes.  In fact now that
I think of it maybe I should stop using the word "mangle".  I only
used it because of C++ "name mangling" but I could call them
PGL_PREFIX_TYPES and PGL_PREFIX_GLSL...something to think about