Fix bug in normalizing non-float attributes
This fixes the problem with my PortableGL IMGUI backend.

Now I just need to figure out how to get window resizing to work and
then clean it up. I haven't decided whether I really want to submit it
to IMGUI or not. For performance, using the sdl2_renderer2 or sdl3_renderer3
backends are far far better but if you were using PGL with something
other than SDL and still wanted IMGUI and performance wasn't an issue...