Replace PGL_EXCLUDE_GLSL with ...
PGL_PREFIX_GLSL or PGL_SUFFIX_GLSL.

I'm not really happy with this solution because it's not very elegant
and feels hacky ... but I wasn't happy with any of the potential
solutions that I've come up with.  Naming is hard enough, naming when
having to factor in other people wanting to use the same names because
you're both implementing the same specification is even harder. Through
in the fact that GLSL, like C++, has function overloading and C does
not... It's enough to drive me mildly crazy because I just want to
make a final decision and move on to more interesting problems.

I also added documentation for PGL_HERMITE_SMOOTHING to the header docs
which I forgot to add last time.